add spin/charge hyperparameters to systemhypers

+++ src/metatrain/utils/data/dataset.py
@@ -1108,7 +1108,10 @@ class MemmapDataset(TorchDataset):
                     keys=Labels.single(),
                     blocks=[
                         TensorBlock(
-                            values=torch.tensor([[float(self.charge_array[i])]]),
+                            values=torch.tensor(
+                                [[float(self.charge_array[i])]],
+                                dtype=torch.float64,
+                            ),
                             samples=Labels(
                                 "system", torch.tensor([[i]], dtype=torch.int32)
                             ),
@@ -1125,7 +1128,10 @@ class MemmapDataset(TorchDataset):
                     keys=Labels.single(),
                     blocks=[
                         TensorBlock(
-                            values=torch.tensor([[float(self.spin_array[i])]]),
+                            values=torch.tensor(
+                                [[float(self.spin_array[i])]],
+                                dtype=torch.float64,
+                            ),
                             samples=Labels(
                                 "system", torch.tensor([[i]], dtype=torch.int32)
                             ),
